API model/requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
pyjwt
python-josepip